+++ lib/archer/history.rb
@@ -1,5 +1,7 @@
 module Archer
   class History < ActiveRecord::Base
     self.table_name = "archer_history"
+
+    serialize :commands, coder: Coder
   end
 end