+++ test/test_helper.rb
@@ -5,6 +5,7 @@ require "minitest/autorun"
 require "minitest/pride"
 require "rails/command"
 require "rails/commands/console/console_command"
+require "irb"
 
 puts "Testing with IRB #{IRB::VERSION}"
 
@@ -12,6 +13,7 @@ logger = ActiveSupport::Logger.new(ENV["VERBOSE"] ? STDOUT : nil)
 
 Combustion.path = "test/internal"
 Combustion.initialize! :active_record do
+  config.load_defaults Rails::VERSION::STRING.to_f
   config.active_record.logger = logger
 end